Improve optimization
The main change is to add a third possible result of folding instructions, which is a no-op. That means that no instructions will be emitted.

Previously, if we added and then subtracted the same amount, or vice versa, we would not be able to optimize that because the Inc and Dec instructions take a non-zero value.
However, really, what we want to do is recognize that that set of instructions would do nothing at all, and remove it entirely.
The new NoOp variant of FoldResult represents that condition.